rearrange the utils folder so that all variables are able to be imported from "/utils"
