+++ utils/index.js
@@ -1,17 +1,4 @@
-import { DEFAULT_WARE_IMAGE } from './constants'
-
-export const configure_services = ({ data, path }) => {
-  return data?.map(ware => {
-    const img = ware.promo_image
-      ? { src: ware.promo_image, alt: `The promotional image for ${ware.name}` }
-      : DEFAULT_WARE_IMAGE
-
-    return {
-      description: ware.snippet,
-      id: ware.reference_of_id,
-      img,
-      name: ware.name,
-      href: `${path}/${ware.slug}`,
-    }
-  })
-}
+export * from './api'
+export * from './configurations'
+export * from './constants'
+export * from './theme'
