+++ utils/theme/index.js
@@ -1,0 +1,2 @@
+export * from './colors'
+export * from './typography'
